move main index file
