+++ src/modules/content-tab-sync/tests/index.test.tsx
@@ -1,14 +1,14 @@
-// To run tests, execute `npm run test -- src/components/tests/content-tab-sync.test.tsx` from the root directory
+// To run tests, execute `npm run test -- src/modules/content-tab-sync/tests/index.test.tsx` from the root directory
 import { render, screen, fireEvent } from '@testing-library/react';
 import escapeRegExp from 'lodash/escapeRegExp';
 import { Provider } from 'react-redux';
-import { ContentTabSync } from 'src/components/content-tab-sync';
 import { SyncSitesProvider, useSyncSites } from 'src/hooks/sync-sites';
 import { SyncPushState } from 'src/hooks/sync-sites/use-sync-push';
 import { useAuth } from 'src/hooks/use-auth';
 import { ContentTabsProvider } from 'src/hooks/use-content-tabs';
 import { useFeatureFlags } from 'src/hooks/use-feature-flags';
 import { getIpcApi } from 'src/lib/get-ipc-api';
+import { ContentTabSync } from 'src/modules/content-tab-sync';
 import { store } from 'src/stores';
 
 jest.mock( 'src/hooks/use-auth' );
